refactor: update test of view model based on the refactor that have done.

+++ viewModel/src/test/java/com/example/viewmodel/utils/entityHelper/createMovie.kt
@@ -16,7 +16,7 @@ fun createMovie(
         name = name,
         description = description,
         posterUrl = poster,
-        productionYear = productionYear,
+        productionYear = productionYear.toUInt(),
         categories = genres,
         rating = rating,
         popularity = 0.0
